requirements.txt
# Deep Learning Frameworks
tensorflow>=2.13.0
keras>=2.13.0
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# IoT and Communication
paho-mqtt>=1.6.1
pyserial>=3.5

# Web Dashboard
flask>=2.3.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# Monitoring and Logging
tensorboard>=2.13.0
wandb>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
